requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-core
langchain-openai
langchain-anthropic
crewai
crewai-tools
pydantic
python-dotenv
httpx
sqlmodel
psycopg2-binary
PyGithub
boto3
